requirements.txt
certifi==2018.4.16
chardet==3.0.4
colorpy==0.1.1
cycler==0.10.0
Django==2.0.5
idna==2.6
image==1.5.20
kiwisolver==1.0.1
matplotlib==2.2.2
mss==3.2.0
numpy==1.14.3
phue==1.0
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
